+++ packages/qb/pqb/src/queryMethods/upsertOrCreate.ts
@@ -5,9 +5,13 @@ import { WhereResult } from './where';
 import { MoreThanOneRowError } from '../errors';
 import { isObjectEmpty } from 'orchid-core';
 
+export type UpsertCreateArg<T extends Query> =
+  | CreateData<T>
+  | (() => CreateData<T>);
+
 export type UpsertData<T extends Query> = {
   update: UpdateData<T>;
-  create: CreateData<T>;
+  create: UpsertCreateArg<T>;
 };
 
 export type UpsertResult<T extends Query> = T['meta']['hasSelect'] extends true
@@ -32,14 +36,14 @@ export class QueryUpsertOrCreate {
 
   orCreate<T extends UpsertThis>(
     this: T,
-    data: CreateData<T>,
+    data: UpsertCreateArg<T>,
   ): UpsertResult<T> {
     return this.clone()._orCreate(data);
   }
 
   _orCreate<T extends UpsertThis>(
     this: T,
-    data: CreateData<T>,
+    data: UpsertCreateArg<T>,
   ): UpsertResult<T> {
     this.query.returnType = 'one';
     this.query.wrapInTransaction = true;
@@ -53,6 +57,10 @@ export class QueryUpsertOrCreate {
 
     this.query.patchResult = async (queryResult) => {
       if (queryResult.rowCount === 0) {
+        if (typeof data === 'function') {
+          data = data();
+        }
+
         const inner = (this as Query).create(data as CreateData<Query>);
         const { handleResult } = inner.query;
         inner.query.handleResult = (q, r, s) => {